add resetControlImage prop for specifying custom icons
