+++ src/components/my-map/controls.ts
@@ -1,6 +1,9 @@
-import "ol/ol.css";
 import { Control, ScaleLine } from "ol/control";
+import "ol/ol.css";
+import eraseIcon from "./icons/erase.svg";
 import northArrowIcon from "./icons/north-arrow-n.svg";
+import resetIcon from "./icons/reset.svg";
+import trashCanIcon from "./icons/trash-can.svg";
 
 export function scaleControl(useScaleBarStyle: boolean) {
   return new ScaleLine({
@@ -20,6 +23,35 @@ export function northArrowControl() {
   const element = document.createElement("div");
   element.className = "north-arrow-control ol-unselectable ol-control";
   element.appendChild(image);
+
+  return new Control({ element: element });
+}
+
+export function resetControl(listener: any, icon: string) {
+  const button = document.createElement("button");
+  button.title = "Reset map view";
+
+  if (icon === "unicode") {
+    button.innerHTML = "↻";
+  } else {
+    const propToSVGLookup: any = {
+      reset: resetIcon,
+      trash: trashCanIcon,
+      erase: eraseIcon,
+    };
+    const image = document.createElement("img");
+    image.className = "reset-icon";
+    image.src = propToSVGLookup[icon];
+    button.appendChild(image);
+  }
+
+  // this is an internal event listener, so doesn't need to be removed later
+  // ref https://lit.dev/docs/components/lifecycle/#disconnectedcallback
+  button.addEventListener("click", listener, false);
+
+  const element = document.createElement("div");
+  element.className = "reset-control ol-unselectable ol-control";
+  element.appendChild(button);
 
   return new Control({ element: element });
 }
