+++ src/components/my-map/icons/README.md
@@ -1,1 +1,1 @@
-Icons are sourced from [Font-GIS](https://viglino.github.io/font-gis/?fg=arrow-o)
+Icons are sourced from [Font-GIS](https://viglino.github.io/font-gis/?fg=arrow-o) and [Carbon Design System](https://carbondesignsystem.com/guidelines/icons/library)
