+++ src/components/my-map/styles.scss
@@ -30,6 +30,10 @@
   top: 114px;
   left: 0.5em;
 }
+.reset-control img {
+  width: 30px;
+  height: auto;
+}
 .north-arrow-control img {
   width: 44px;
   height: auto;
